user role permission added with tested

+++ resources/views/users/index.blade.php
@@ -2,11 +2,13 @@
 @section('content')
     <div class="row">
         <div class="col-lg-12 margin-tb">
-            <div class="pull-left">
+            <div class="float-left">
                 <h2>Users Management</h2>
             </div>
-            <div class="pull-right">
-                <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a>
+            <div class="float-right">
+                @can('user-create')
+                    <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a>
+                @endcan
             </div>
         </div>
     </div>
@@ -37,10 +39,14 @@
                 </td>
                 <td>
                     <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
-                    <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
-                    {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
-                    {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-                    {!! Form::close() !!}
+                    @can('user-edit')
+                        <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+                    @endcan
+                    @can('user-delete')
+                        {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                        <button type="submit" class="btn btn-danger" onclick="return confirm('Are you sure?')">Delete</button>
+                        {!! Form::close() !!}
+                    @endcan
                 </td>
             </tr>
         @endforeach
